kexec: add KHO support to kexec file loads
Kexec has 2 modes: A user space driven mode and a kernel driven mode.
For the kernel driven mode, kernel code determines the physical
addresses of all target buffers that the payload gets copied into.

With KHO, we can only safely copy payloads into the "scratch area".
Teach the kexec file loader about it, so it only allocates for that
area. In addition, enlighten it with support to ask the KHO subsystem
for its respective payloads to copy into target memory. Also teach the
KHO subsystem how to fill the images for file loads.

Signed-off-by: Alexander Graf <graf@amazon.com>
Co-developed-by: Mike Rapoport (Microsoft) <rppt@kernel.org>
Signed-off-by: Mike Rapoport (Microsoft) <rppt@kernel.org>
Co-developed-by: Changyuan Lyu <changyuanl@google.com>
Signed-off-by: Changyuan Lyu <changyuanl@google.com>

kexec: defer KHO FDT finalization until kernel_kexec()

Previously the KHO FDT must be created before kexec_file_load(), and
after that no new data can be added to the FDT, which essentially makes
kexec_file_load() part of the blackout window of kernel kexec.

In this patch, kho_fill_kimage() reserves an empty segment in kimage
while the actual FDT data is copied to the segment pages in
kernel_kexec(). Additionally, if the FDT is not created yet,
kernel_kexec() notifies KHO users to finalize the FDT blob.

In this way, the FDT finalization and kexec_file_load() are fully
decoupled.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>