kexec: defer KHO FDT finalization until kernel_kexec()
Previously the KHO FDT must be created before kexec_file_load(), and
after that no new data can be added to the FDT, which essentially makes
kexec_file_load() part of the blackout window of kernel kexec.

In this patch, kho_fill_kimage() reserves an empty segment in kimage
while the actual FDT data is copied to the segment pages in
kernel_kexec(). Additionally, if the FDT is not created yet,
kernel_kexec() notifies KHO users to finalize the FDT blob.

In this way, the FDT finalization and kexec_file_load() are fully
decoupled.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>